fix image panel circular dependencies
